move components to commoon
