+++ src/App.vue
@@ -60,12 +60,12 @@ import {
 } from '@/state/app'
 import { CHECK_FOR_UPDATES } from '@/messages'
 
-import TitleBar from './components/TitleBar.vue'
-import Navbar from './components/Navbar/Navbar.vue'
-import PlayerContainer from './components/Player/Container.vue'
-import ToastOverlay from './components/ToastOverlay.vue'
-import AboutModal from './components/Modals/AboutModal.vue'
-import EditModal from './components/Modals/EditModal.vue'
+import TitleBar from '@/common/components/title-bar.vue'
+import Navbar from '@/components/Navbar/Navbar.vue'
+import PlayerContainer from '@/components/Player/Container.vue'
+import ToastOverlay from '@/components/ToastOverlay.vue'
+import AboutModal from '@/components/Modals/AboutModal.vue'
+import EditModal from '@/components/Modals/EditModal.vue'
 import ManualSearchModal from '@/components/Modals/ManualSearch/ManualSearchModal.vue'
 import { Hidive } from '@/lib/hidive'
