+++ src/common/components/animated-height.vue
