+++ src/common/components/episode-list.vue
@@ -38,9 +38,9 @@ import { Required } from '@/decorators'
 import { Hidive, HidiveResponseCode } from '@/lib/hidive'
 import { isNil } from '@/utils'
 
-import Episode from './Episode.vue'
-import Loading from './Loading.vue'
-import SourceList from './SourceList.vue'
+import Episode from './episode.vue'
+import Loading from './loading.vue'
+import SourceList from './source-list.vue'
 
 @Component({ components: { SourceList, Loading, Episode } })
 export default class EpisodeList extends Vue {
@@ -133,7 +133,7 @@ export default class EpisodeList extends Vue {
 </script>
 
 <style scoped lang="scss">
-@import '../colors';
+@import '../../colors';
 
 .loading-wrapper {
   padding: 10px;
