+++ src/common/components/form/dropdown.vue
@@ -43,7 +43,7 @@ export default class Dropdown extends Vue {
 </script>
 
 <style scoped lang="scss">
-@import '../../colors';
+@import '../../../colors';
 
 .dropdown {
   position: relative;
