+++ src/common/components/form/number-input.vue
@@ -67,7 +67,7 @@ export default class NumberInput extends Vue {
 </script>
 
 <style scoped lang="scss">
-@import '../../colors';
+@import '../../../colors';
 
 .number-input {
   position: relative;
