+++ src/common/components/form/text-input.vue
@@ -39,7 +39,7 @@ export default class TextInput extends Vue {
 </script>
 
 <style scoped lang="scss">
-@import '../../colors';
+@import '../../../colors';
 
 .text-input {
   position: relative;
