+++ src/views/first-time-setup/components/connections.vue
@@ -63,7 +63,7 @@ import crIcon from '@/assets/crunchyroll.svg'
 import hidiveIcon from '@/assets/hidive.svg'
 import LoginCr from '@/common/components/login/crunchyroll.vue'
 import LoginHd from '@/common/components/login/hidive.vue'
-import TextInput from '@/components/Form/TextInput.vue'
+import TextInput from '@/common/components/form/text-input.vue'
 import CButton from '@/components/CButton.vue'
 
 import { getFinishedConnecting, getIsConnectedTo } from '@/state/auth'
